+++ UnitTests/BraintreePayPalTests/BTPayPalVaultRequest_Tests.swift
@@ -93,6 +93,20 @@ class BTPayPalVaultRequest_Tests: XCTestCase {
         XCTAssertEqual(parameters["merchant_app_return_url"] as? String, "some-url")
     }
     
+    func testParametersWithConfiguration_setsAppSwitchParameters_withoutUserAuthenticationEmail() {
+        let request = BTPayPalVaultRequest(
+            userAuthenticationEmail: "",
+            enablePayPalAppSwitch: true
+        )
+        
+        let parameters = request.parameters(with: configuration, universalLink: URL(string: "some-merchant-url")!, isPayPalAppInstalled: true)
+        
+        XCTAssertEqual(parameters["launch_paypal_app"] as? Bool, true)
+        XCTAssertTrue((parameters["os_version"] as! String).matches("\\d+\\.\\d+"))
+        XCTAssertTrue((parameters["os_type"] as! String).matches("iOS|iPadOS"))
+        XCTAssertEqual(parameters["merchant_app_return_url"] as? String, "some-merchant-url")
+    }
+    
     func testParameters_withRecurringBillingDetails_returnsAllParams() {
         let billingPricing = BTPayPalBillingPricing(
             pricingModel: .autoReload,