+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unreleased
 * BraintreeShopperInsights (BETA)
   * Bug fix: remove internal code signing from framework
+* BraintreePayPal
+  * Make `userAuthenticationEmail` property optional for all App Switch flows
 
 ## 6.30.0 (2025-03-17)
 * Attach zip file of static Braintree xcframework modules to each GitHub release